+++ magda-sleuther-framework/src/setupWebhookEndpoint.ts
@@ -6,22 +6,16 @@ import Registry from "@magda/typescript-common/dist/registry/AuthorizedRegistryC
 import unionToThrowable from "@magda/typescript-common/dist/util/unionToThrowable";
 
 import SleutherOptions from "./SleutherOptions";
-import getWebhookUrl from "./getWebhookUrl";
+
 import AsyncPage, {
     forEachAsync
 } from "@magda/typescript-common/dist/AsyncPage";
 
 export default function setupWebhookEndpoint(
+    server: express.Application,
     options: SleutherOptions,
     registry: Registry
 ) {
-    const server = options.express();
-    server.use(require("body-parser").json({ limit: "50mb" }));
-
-    server.get("/healthz", (request, response) => {
-        response.status(200).send("OK");
-    });
-
     server.post(
         "/hook",
         (request: express.Request, response: express.Response) => {
@@ -118,11 +112,4 @@ export default function setupWebhookEndpoint(
             }
         }
     );
-
-    function getPort() {
-        return options.argv.listenPort;
-    }
-
-    server.listen(getPort());
-    console.info(`Listening at ${getWebhookUrl(options)}`);
 }